make secureSession method inside the XenIssuing class
